extract declartation emitter to a separate file
