+++ src/compiler/types.ts
@@ -1781,4 +1781,20 @@ module ts {
         // of this method before/after the operation is performed.
         getModificationCount(): number;
     }
+
+    // @internal
+    export interface EmitTextWriter {
+        write(s: string): void;
+        writeTextOfNode(sourceFile: SourceFile, node: Node): void;
+        writeLine(): void;
+        increaseIndent(): void;
+        decreaseIndent(): void;
+        getText(): string;
+        rawWrite(s: string): void;
+        writeLiteral(s: string): void;
+        getTextPos(): number;
+        getLine(): number;
+        getColumn(): number;
+        getIndent(): number;
+    }
 }
